+++ common/src/addrtranslate-auxv.C
@@ -28,7 +28,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
  */
 
-#include "common/src/parseauxv.h"
+#include "common/src/AuxvParser.h"
 #include "common/src/headers.h"
 
 #include "common/src/addrtranslate.h"
