+++ common/src/addrtranslate-sysv.C
@@ -47,7 +47,7 @@
 
 #include <iostream>
 
-#include "common/src/parseauxv.h"
+#include "common/src/AuxvParser.h"
 #include "common/src/headers.h"
 #include "common/src/dyninst_filesystem.h"
 #include "common/src/vm_maps.h"
