+++ common/src/linuxKludges.C
@@ -29,7 +29,7 @@
  */
 
 #include "common/src/headers.h"
-#include "common/src/parseauxv.h"
+#include "common/src/AuxvParser.h"
 #include "common/src/linuxKludges.h"
 #include <elf.h>
