+++ proccontrol/src/linux.C
@@ -71,7 +71,7 @@
 
 #include "common/src/linuxKludges.h"
 #include "linuxHeaders.h"
-#include "common/src/parseauxv.h"
+#include "common/src/AuxvParser.h"
 
 #include "boost/shared_ptr.hpp"
